修复管理员认证

+++ src/main/java/cn/edu/upc/yb/integrate/calendar/controller/CalendarController.java
@@ -26,13 +26,15 @@ public class CalendarController {
 
     @RequestMapping("/create")
     public Object creatCalendar(String schoolschedule, String begindate, String enddate) {
+        if (commonAdminService.isCommonAdmin() == false) return new ErrorReporter(-1, "您没有权限操作");
         SchoolCalendar schoolCalendar = new SchoolCalendar(schoolschedule, begindate, enddate);
         schoolCalendarDao.save(schoolCalendar);
         return new JsonMes(1, "创建成功");
     }
 
     @RequestMapping("/update")
-    public Object updateCalendar(@RequestParam(value = "id",defaultValue = "0") int id, String schoolschedule, String begindate, String enddate) {
+    public Object updateCalendar(@RequestParam(value = "id", defaultValue = "0") int id, String schoolschedule, String begindate, String enddate) {
+        if (commonAdminService.isCommonAdmin() == false) return new ErrorReporter(-1, "您没有权限操作");
         SchoolCalendar schoolCalendar = schoolCalendarDao.findOne(id);
         schoolCalendar.updata(schoolschedule, begindate, enddate);
         schoolCalendarDao.save(schoolCalendar);
@@ -40,7 +42,8 @@ public class CalendarController {
     }
 
     @RequestMapping("/delete")
-    public Object deleteCalendar(@RequestParam(value = "id",defaultValue = "0") int id) {
+    public Object deleteCalendar(@RequestParam(value = "id", defaultValue = "0") int id) {
+        if (commonAdminService.isCommonAdmin() == false) return new ErrorReporter(-1, "您没有权限操作");
         SchoolCalendar schoolCalendar = schoolCalendarDao.findOne(id);
         schoolCalendar.delete();
         schoolCalendarDao.save(schoolCalendar);
@@ -49,6 +52,7 @@ public class CalendarController {
 
     @RequestMapping("/showcalendar")
     public Object showCalendar(String schoolschedule) {
+        if (commonAdminService.isCommonAdmin() == false) return new ErrorReporter(-1, "您没有权限操作");
         return schoolCalendarDao.findBySchoolscheduleAndIsdelete(schoolschedule, false);
     }
 
@@ -56,8 +60,8 @@ public class CalendarController {
      * 管理员显示全部
      */
     @RequestMapping("/showall")
-    public Object showAll(){
-        if(commonAdminService.isCommonAdmin() == false) return new ErrorReporter(-1,"您没有权限操作");
+    public Object showAll() {
+        if (commonAdminService.isCommonAdmin() == false) return new ErrorReporter(-1, "您没有权限操作");
         return schoolCalendarDao.findByIsdeleteOrderByIdDesc(false);
     }
